Feature/calculate backlog proyection (#827)
* add backlog projections

* add mapping to projection response

* test is added and the code is organized

* fix code quality issues

* fix code quality issues

* merge with develop

* fix cq checks

* add code-quality configuration

* pr comments

* fix test

* refactor

* fix cq issues

* refactor and fix method that map waving and picking backlog

* pr comments

* cq checks

+++ src/main/java/com/mercadolibre/planning/model/api/projection/dto/request/BacklogProjection.java
@@ -3,7 +3,7 @@ package com.mercadolibre.planning.model.api.projection.dto.request;
 import static java.util.stream.Collectors.flatMapping;
 import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.mapping;
-import static java.util.stream.Collectors.summingLong;
+import static java.util.stream.Collectors.summingInt;
 import static java.util.stream.Collectors.toMap;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -57,14 +57,14 @@ public class BacklogProjection {
             ));
     }
 
-    public Map<ProcessName, Map<Instant, Long>> mapThroughput() {
+    public Map<ProcessName, Map<Instant, Integer>> mapThroughput() {
       return this.throughput.stream().flatMap(tph -> tph.getQuantityByProcessName().stream()
               .map(quantityByProcessName -> new AbstractMap.SimpleEntry<>(tph.getOperationHour(), quantityByProcessName)))
           .collect(groupingBy(
               a -> a.getValue().getName(),
               groupingBy(
                   AbstractMap.SimpleEntry::getKey,
-                  summingLong(entry -> (long) entry.getValue().getTotal())
+                  summingInt(entry -> entry.getValue().getTotal())
               )
           ));
     }
